requirements.txt
python-vagrant==0.5.15
fabric==2.6.0
requests==2.27.1
simplejson==3.17.6
tox==3.24.5
coverage==6.3.2
PyYAML==6.0
straight.plugin==1.5.0
paramiko==2.10.1
pymetasploit3==1.0.3
pylint==2.12.2
flask==2.0.3
pydantic
dotmap==1.3.26
eel

# Sphinx stuff
sphinx>=4.2.0
sphinx-argparse
sphinxcontrib-autoyaml
sphinx-pyreverse>=0.0.17
sphinxcontrib.asciinema>=0.3.3
sphinx-revealjs
# sphinx-pydantic    # This one has issues that must be fixed upstream first

# Mypy stuff
mypy==0.931
types-PyYAML==6.0.4
types-requests==2.27.11
types-simplejson==3.17.3
types-paramiko==2.8.13

# pipdeptree
pipdeptree
pip-upgrade

# Argcomplete. See README.md
argcomplete==2.0.0
